refactor: move Keycloak configuration to auth package
